+++ src/main/kotlin/assertk/assertions/list.kt
@@ -1,9 +1,7 @@
 package assertk.assertions
 
 import assertk.Assert
-import assertk.assertions.support.expected
-import assertk.assertions.support.fail
-import assertk.assertions.support.show
+import assertk.assertions.support.*
 
 /**
  * Returns an assert that assertion on the value at the given index in the list.
@@ -20,3 +18,22 @@ fun <T> Assert<List<T>>.index(index: Int, f: (Assert<T>) -> Unit) {
     }
 }
 
+/**
+ * Asserts the list contains exactly the expected elements. They must be in the same order and
+ * there must not be any extra elements.
+ * @see [containsAll]
+ */
+fun <T : List<*>> Assert<T>.containsExactly(vararg elements: Any?) {
+    if (actual == elements.asList()) return
+
+    val diff = ListDiffer.diff(elements.asList(), actual)
+        .filterNot { it is ListDiffer.Edit.Eq }
+
+    expected(diff.joinToString(prefix = "to contain exactly:\n", separator = "\n") { edit ->
+        when (edit) {
+            is ListDiffer.Edit.Del -> " at index:${edit.oldIndex} expected:${show(edit.oldValue)}"
+            is ListDiffer.Edit.Ins -> " at index:${edit.newIndex} unexpected:${show(edit.newValue)}"
+            else -> throw IllegalStateException()
+        }
+    })
+}